requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
xarray>=2023.1.0
netCDF4>=1.6.0
dask>=2023.1.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
chromadb>=0.4.0
faiss-cpu>=1.7.0

# AI/ML
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.0
openai>=1.0.0

# Visualization
plotly>=5.15.0
folium>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data processing
pyarrow>=12.0.0
fastparquet>=0.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
